requirements.txt
Django==2.0.7
django-filter==1.1.0
djangorestframework==3.8.2
drfaddons==0.0.1
pkg-resources==0.0.0
pytz==2018.5
